backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
email-validator==2.2.0
cryptography==43.0.3
python-multipart==0.0.12
itsdangerous==2.2.0
orjson==3.10.7
structlog==24.4.0
aiofiles==24.1.0
slowapi==0.1.9
bleach==6.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
gunicorn==23.0.0


